companion/requirements.txt
aiohttp>=3.9
websockets>=12.0
Pillow>=10.0
